feat: refactor new type to common location
